pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nfckd"
version = "0.1.0"
description = "A Python library for NFC tag writing and key derivation"
readme = "README.md"
authors = [
    { name = "Zears14"}
]
license = { text = "MIT" }
dependencies = [
    "cryptography>=45.0.5",
    "nfcpy>=1.0.4",
    "loguru>=0.7.3"
]
[tool.setuptools]
packages = ["nfckd"]

